% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\name{ArrangeTypeIerror}
\alias{ArrangeTypeIerror}
\title{Uniform QQ-plot.}
\usage{
ArrangeTypeIerror(path, case, textsize = c(20, 20, 8, 20), showcol = c(1,
  4:9), legend = F, color1 = "#0099FF", color2 = c("#000000", "#E69F00",
  "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00"))
}
\arguments{
\item{case}{which correlation structure to plot.}

\item{textsize, }{a vector of four, controlling the text size of legened, title, axis, axis.title, respectively}

\item{legend}{if \code{TRUE}, plot the legend; otherwise just plot the qq-plot}

\item{dat1}{simulated p-value matrix when DE % = 0.}

\item{dat2}{simulated p-value matrix for DE % = 10%.}
}
\value{
a figure
}
\description{
Combines two qq plot in one figure and produce the legend in a separate figure.
}

