% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaca.R
\name{btw_gene_corr}
\alias{btw_gene_corr}
\title{Estimate sample correlation.}
\usage{
btw_gene_corr(expression_data, trt, geneset, standardize = T,
  minSetSize = 5)
}
\arguments{
\item{expression_data}{the expressoin matrix.}

\item{trt}{treatment labels}

\item{geneset}{an object from \code{read_gene_set}}

\item{standardize}{`TRUE` or `FALSE`, whether the data should be standaridzed}

\item{minSetSize}{the minimum number of genes contained for a gene set to be
considered.}
}
\value{
a list 
\item{set_name}{The name of the gene set}
\item{testSetCor}{Average correlation for genes in the test set}
\item{interCor}{Average correlation between genes in the test set and those not in the test set}
\item{backSetCor}{Average correlations for genes not in the test set.}
}
\description{
Average correlations for genes
}

