% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE64810.R
\name{Camera_multiple1}
\alias{Camera_multiple1}
\title{CAMERA multiple}
\usage{
Camera_multiple1(expression, trt, geneset, use.rank = F)
}
\description{
multiple gene set enrichemnt analysis (specific for testing GSE64810)
}

