% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{compare_test}
\alias{compare_test}
\title{Compare meaca to existing methods}
\usage{
compare_test(dat)
}
\arguments{
\item{dat}{the simulated data consisting of three lists:
\item{data}{the expression matrix}
\item{trt}{what treatment each column of \code{data} belongs to}
\item{go_term}{whether each row of \code{data} belongs to this go term}}
}
\value{
a matrix of p values
}
\description{
Compare meaca to other existing methods.
}
\keyword{internal}

