% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\name{createEmpiricalPowerTable}
\alias{createEmpiricalPowerTable}
\title{create the emperical power table .}
\usage{
createEmpiricalPowerTable(path, showcol, setup = "A1", case = "a")
}
\arguments{
\item{path}{where the simulated p-value matrix locates.}

\item{showcol}{for which methods power will be summarized}

\item{setup}{if it's group I (\code{A1}) or group II (\code{A2}) simulation.}
}
\value{
a table
}
\description{
the table in the emperical power simulation
}

