% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{data_simu_meaca}
\alias{data_simu_meaca}
\title{Power for meaca}
\usage{
data_simu_meaca(case, prop = c(0.1, 0.1), de = c(2, 1), nsim = 1000,
  size = 50, rho = c(0.1, 0.05, -0.05),
  dest = "/home/stats/zhuob/data/computing/", num_gene = c(500, 100))
}
\arguments{
\item{prop}{A vector of two, specifying the proportion of DE genes for the test set and the background set}

\item{de}{A vector of two, specifying the (normally distributed) DE effect size and its std}

\item{nsim}{number of simulations to run}

\item{size}{number of samples to be simulated}

\item{rho}{A vector of three, for correlation coefficients corresponding to rho1, rho2, rho3 in the paper}

\item{dest}{where to store the results}

\item{num_gene}{A vector of two, specifying the number of genes to be simulated in total and in the test set}

\item{post_txt}{formated name of output files}
}
\value{
a text file containing the p value matrix
}
\description{
Produce p value matrix for simulation discussed in the paper (Figure 2).
}

