% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportive-functions.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate sample covariance.}
\usage{
estimate_sigma(expression_data, trt)
}
\arguments{
\item{expression_data}{the expression matrix.}

\item{trt}{sample labels. 0 for control and 1 for treatment}
}
\value{
a list 
\item{sigma}{a covariance matrix }
\item{t_val}{a vector of gene level test statistics}
}
\description{
Estimate sample covariance and calculate the gene-level statistics
}
