% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.1.0.R, R/read-gene-set.R
\name{GSEA.ReadClsFile}
\alias{GSEA.ReadClsFile}
\title{readling Cls files.}
\usage{
GSEA.ReadClsFile(file = "NULL")

GSEA.ReadClsFile(file = "NULL")
}
\arguments{
\item{msigdb}{gene set ensemble downloaded from broad institute.}
}
\value{
a list 
\item{total} { number of gene sets contained.}
\item{size} {a numerical vector containing the size of each gene set.}
\item{gene_set} { a list. The first element is the set name. From the third
element each containing members of the gene set.}
}
\description{
readling Cls file (from GSEA.1.0.R)

read the gene sets of the MsigDB format.
}
\keyword{internal}

