% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE64810.R
\name{GSEA1}
\alias{GSEA1}
\title{GSEA multiple}
\usage{
GSEA1(input.ds, input.cls, gene.ann = "", gs.db, gs.ann = "",
  output.directory = "", doc.string = "GSEA.analysis",
  non.interactive.run = F, reshuffling.type = "sample.labels",
  nperm = 1000, weighted.score.type = 1, nom.p.val.threshold = -1,
  fwer.p.val.threshold = -1, fdr.q.val.threshold = 0.25, topgs = 10,
  adjust.FDR.q.val = F, gs.size.threshold.min = 25,
  gs.size.threshold.max = 500, reverse.sign = F, preproc.type = 0,
  random.seed = 123456, perm.type = 0, fraction = 1, replace = F,
  save.intermediate.results = F, OLD.GSEA = F,
  use.fast.enrichment.routine = T)
}
\description{
multiple gene set enrichemnt analysis (specific for testing GSE64810)
}

