% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE64810.R
\name{meaca_multiple1}
\alias{meaca_multiple1}
\title{meaca multiple}
\usage{
meaca_multiple1(expression_data1, trt, geneset, standardize = T,
  minSetSize = 5, fdr_method = "BH")
}
\description{
multiple gene set enrichemnt analysis (specific for testing GSE64810)
}

