% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{MRGSE_multiple}
\alias{MRGSE_multiple}
\title{MRGSE multiple}
\usage{
MRGSE_multiple(expression, trt, geneset, use.rank = TRUE)
}
\arguments{
\item{expression}{the expression data matrix}

\item{trt}{the treatment lables for each column of \code{expression_data}}

\item{geneset}{a list containing all the gene sets to be tested for enrichment status}

\item{use.rank}{If \code{TRUE} then it corresponds to "MRGSE", otherwise "Camara"}
}
\value{
a matrix
}
\description{
multiple gene set enrichemnt analysis
}

