% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE64810.R
\name{MRGSE_multiple1}
\alias{MRGSE_multiple1}
\title{MRGSE multiple}
\usage{
MRGSE_multiple1(expression, trt, geneset, use.rank = T)
}
\description{
multiple gene set enrichemnt analysis (specific for testing GSE64810)
}

