% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\name{plot_fig1}
\alias{plot_fig1}
\title{Uniform QQ-plot.}
\usage{
plot_fig1(path, textsize = c(15, 15, 10, 10), case = letters[1:5],
  showcol = c(1, 4:9), color1 = "#0099FF", color2 = c("#000000",
  "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00"))
}
\arguments{
\item{path}{where the simulated p-value matrix locates.}

\item{textsize, }{a vector of four, controlling the text size of legened, title, axis, axis.title, respectively}

\item{case}{which correlation structures will show}

\item{showcol}{the methods chosen to compare}

\item{color1}{Color for the reference line}

\item{color2}{Colors for different methods. It must have the same length as the that of \code{showcol}}
}
\value{
a figure
}
\description{
uniform qq-plot for type I error simulation.
}

