% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\name{plot_fig2}
\alias{plot_fig2}
\title{Power plot with 95% error bar.}
\usage{
plot_fig2(path, showcol = 1, textsize = rep(20, 4),
  xtext = c("DE percentage", "Power"))
}
\arguments{
\item{path}{where the simulated p-value matrix locates.}

\item{showcol}{which method to summarize (1 for "MEACA").}

\item{textsize}{a vector of four, controlling the text size of legened, title, axis, axis.title, respectively}

\item{xtext}{the title for x-axis and y-axis}
}
\value{
a figure
}
\description{
Plot figure 2 of the paper.
}

