% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-expression.R
\name{prepare_go_term}
\alias{prepare_go_term}
\title{Simulate gene labels.}
\usage{
prepare_go_term(num_gene, prop, delta)
}
\arguments{
\item{num_gene}{a vector of length 2, total number of genes to be simulated and number of genes in the test set.}

\item{prop}{a vector of length 2, proportion of DE genes within go term and outside go_term, corresponding to $p_t$ and $p_b$.}

\item{delta}{a vector of length num_gene, representing DE effect for each gene}
}
\value{
a list
\item{go_term}{ DE indicator for each simulated gene. 1 is for genes in the test set, and 0 othwerwise.}
\item{delta}{DE effect for each simulated gene.}
\item{n_gene}{ number of genes to be simulated.}
}
\description{
Simulate test set gene labels
}

