% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-expression.R
\name{prepare_simulation}
\alias{prepare_simulation}
\title{Simulate expression data}
\usage{
prepare_simulation(rho1, rho2, rho3, ...)
}
\arguments{
\item{rho1}{a scalar, correlation between two test genes (i.e., \eqn{\rho_1} in the paper)}

\item{rho2}{a scalar, correlation between two background genes (i.e., \eqn{\rho_2} in the paper)}

\item{rho3}{correlation between a test gene and a background gene (i.e., \eqn{\rho_3} in the paper)}

\item{...}{other parameters inherited from \code{prepare_go_term}}
}
\value{
a list 
\item{go_term}{ DE indicator for each simulated gene. 1 is for genes in the
test set, and 0 othwerwise}
\item{delta}{DE effect for each simulated gene}
\item{sigma}{the true correlation, used to simulate  the expression data}
}
\description{
Prepare simulation parameters.
}
\keyword{internal}

