% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{sig_path}
\alias{sig_path}
\title{Sigpathway}
\usage{
sig_path(index, statistics, nsim = 9999)
}
\arguments{
\item{num_gene}{number of genes to be simulated.}
}
\value{
p   the p-values.
}
\description{
implement the Sigpathay.
}

