% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-expression.R
\name{simulate_expression_data}
\alias{simulate_expression_data}
\title{Simulate expression data.}
\usage{
simulate_expression_data(size, obj)
}
\arguments{
\item{size}{number of samples to be simulated}

\item{obj}{object returned by \link{\code{prepare_simulation}}.}
}
\value{
a list 
\item{data}{a expression matrix of \eqn{m\times n} where m is the number of genes and n is the number of samples. }
\item{trt}{sample labels of length \code{n}, 1 for treatment and 0 for control.}
\item{go_term}{gene labels of length \code{m}, 1 for go_term genes and 0 otherwise.}
\item{sigma}{true covariance matrix upon which data is simulated.}
}
\description{
simulate normally distributed expression data with desired DE probabilities for genes in the test set and for those not in the test set..
}

