% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaca.R
\name{standardize_expression_data}
\alias{standardize_expression_data}
\title{standardize expression data, with method described in the paper.}
\usage{
standardize_expression_data(expression_data, trt)
}
\arguments{
\item{expression_data}{the expression matrix.}

\item{trt}{sample labels. 0 for control and 1 for treatment}
}
\value{
a matrix of the same dimension as input data.
}
\description{
Standardize the expression data.
}

