% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\name{summarizePower}
\alias{summarizePower}
\title{summerize power.}
\usage{
summarizePower(path, setup, showcol, case)
}
\arguments{
\item{path}{where the simulated p-value matrix locates.}

\item{setup}{if it's group I (\code{A1}) or group II (\code{A2}) simulation.}

\item{showcol}{which methods to summarize.}

\item{case}{the correlation structure.}
}
\value{
a table
}
\description{
Summarize the power
}

