% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-rnaseq.R
\name{transform_count_edgeR}
\alias{transform_count_edgeR}
\title{Perform count matrix transformation using edgeR procedure}
\usage{
transform_count_edgeR(y, group)
}
\arguments{
\item{y}{the expression count matrix, columns being samples, rows being genes}

\item{group}{a vector of treatment label (e.g., 0 for control, 1 for treatment).}
}
\value{
a matrix of transformed data
}
\description{
Perform count matrix transformation using edgeR procedure
}
\examples{
mu <- matrix(10, 100, 4)
group <- factor(c(0,0,1,1))
design <- model.matrix(~group)
set.seed(123)
library(edgeR)
y0 <- matrix(rnbinom(100*4, mu=mu, size=10),100,4)
y <- DGEList(counts=y0, group=group)
y <- estimateDisp(y, design)

iset1 <- 1:10
camera.DGEList(y, iset1, design)

# the Pvalue should be the same 
y2 <- transform_count_edgeR(y = y0, group = group)
camera(y2, iset1, design)

}
\seealso{
[edgeR::camera.DGEList()]
}
