% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-rnaseq.R
\name{transform_count_vst}
\alias{transform_count_vst}
\title{Perform variance stabilizing transformation for the count matrix data}
\usage{
transform_count_vst(y, group, ...)
}
\arguments{
\item{y}{the expression count matrix, columns being samples, rows being genes}

\item{group}{a vector of treatment label (e.g., 0 for control, 1 for treatment).}

\item{...}{other parameters used in \code{\link[DESeq2]{varianceStabilizingTransformation}}}
}
\value{
a matrix of transformed data
}
\description{
Perform variance stabilizing transformation for the count matrix data
}
\details{
It is absolutely critical that the columns of the count matrix and
  the rows of the column data (information about samples) are in the same
  order. For more details, see \url{https://bioconductor.org/packages/release/bioc/vignettes/DESeq2/inst/doc/DESeq2.html}
}
\examples{
library(DESeq2)
set.seed(123)
y <- matrix(rbinom(6000, 20, 0.4), nrow = 1000)
trt <- c(0, 0, 0, 1, 1, 1)
yr <- transform_count_vst(y = y, group = trt)
go_term <- rep(c(1, 0), c(100, 900)) # first 100 genes in the test set
meaca_single(expression_data = yr, trt = trt, go_term = go_term, standardize = TRUE)
}
\seealso{
\code{\link[DESeq2]{varianceStabilizingTransformation}}
}
