% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-results.R
\name{typeIerror_quantile}
\alias{typeIerror_quantile}
\title{calculate the actual type I error for a nominal alpha}
\usage{
typeIerror_quantile(data, alpha, showcol = 1)
}
\arguments{
\item{data}{simulated p-value matrix.}

\item{alpha, }{a vector or a numeric value that specifies the desired type I error}
}
\value{
the adjusted alpha value.
}
\description{
the recalibrated power
}

