% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportive-functions.R
\name{btw_gene_corr}
\alias{btw_gene_corr}
\title{Calculate sample (Pearson) correlations among gene clusters}
\usage{
btw_gene_corr(expression_data, trt, go_term, standardize = T)
}
\arguments{
\item{expression_data}{the expressoin matrix.}

\item{trt}{treatment indicators, 1 for treatment, 0 for control group}

\item{go_term}{an indicator vector. 1 for genes in the test set, 0 otherwise}

\item{standardize}{whether the data should be standaridzed}
}
\value{
a $1 \times 3$ data frame containing values for $\rho_1$, $\rho_3$
  and $\rho_2$ respectively.
\item{testSetCor}{Average correlation for genes in the test set}
\item{interCor}{Average correlation between genes in the test set and those not in the test set}
\item{backSetCor}{Average correlations for genes not in the test set.}
}
\description{
Calculate sample (Pearson) correlations among gene clusters
}
