% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-wrapper.R
\name{data_simu}
\alias{data_simu}
\title{Compare meaca to existing methods}
\usage{
data_simu(nsim = 1000, ncore = 6, package_used = c("MASS", "qusage"),
  verbose_show = FALSE, meaca_only = FALSE,
  file_to_source = "/home/stats/zhuob/Rcode/Enrichment/GSEA.1.0.R",
  dest = "/home/stats/zhuob/data/computing/", n_gene = 500, n_test = 100,
  prop = c(0.1, 0.1), rho1 = 0.1, rho2 = 0.05, rho3 = -0.05,
  case = "e", size = 50, de_mu = 2, de_sd = 1,
  data_gen_method = "chol", seed = 123)
}
\arguments{
\item{nsim}{number of simulation to run}

\item{ncore}{number of CPUs to be used in the parallel simulation}

\item{package_used}{the packages to be used in the simulation}

\item{verbose_show}{for debug purpose, set to `FALSE` if not in debug mode}

\item{meaca_only}{Should all the methods to be compared? If `TRUE`, produce
Figure 1; otherwise Figure 2}

\item{file_to_source}{the R files containing functions to be sourced}

\item{dest}{where to store the results}

\item{n_gene}{total number of genes to be simulated}

\item{n_test}{number of genes in the test set.}

\item{prop}{a vector of length 2, proportion of DE genes within go term and
outside go_term, corresponding to $p_t$ and $p_b$.}

\item{rho1}{a scalar, correlation between two test genes (i.e., \eqn{\rho_1} in the paper)}

\item{rho2}{a scalar, correlation between two background genes (i.e., \eqn{\rho_2} in the paper)}

\item{rho3}{correlation between a test gene and a background gene (i.e., \eqn{\rho_3} in the paper)}

\item{size}{number of samples to be simulated}

\item{de_mu, de_sd}{if the gene is DE, delta ~ N(de_mu, de_sd)}

\item{data_gen_method}{data generation method; if `data_gen_method = MASS`,
then `MASS::mvrnorm` is used, otherwise see function \link[mvtnorm]{rmvnorm}}

\item{seed}{the seed used for simulation (for reproducibility purpose)}

\item{seed}{the seed used for simulation (for reproducibility purpose)}
}
\value{
a text file containing the p value matrix
}
\description{
Produce p value matrix for simulation discussed in the paper.
}

