% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-expression.R
\name{simulate_expression_data}
\alias{simulate_expression_data}
\title{Simulate expression data.}
\usage{
simulate_expression_data(size, n_gene, n_test, prop, de_mu, de_sd, rho1, rho2,
  rho3, data_gen_method = "chol", seed = 123)
}
\arguments{
\item{size}{number of samples to be simulated}

\item{n_gene}{total number of genes to be simulated}

\item{n_test}{number of genes in the test set.}

\item{prop}{a vector of length 2, proportion of DE genes within go term and
outside go_term, corresponding to $p_t$ and $p_b$.}

\item{de_mu, de_sd}{if the gene is DE, delta ~ N(de_mu, de_sd)}

\item{rho1}{a scalar, correlation between two test genes (i.e., \eqn{\rho_1} in the paper)}

\item{rho2}{a scalar, correlation between two background genes (i.e., \eqn{\rho_2} in the paper)}

\item{rho3}{correlation between a test gene and a background gene (i.e., \eqn{\rho_3} in the paper)}

\item{data_gen_method}{data generation method; if `data_gen_method = MASS`,
then \link[MASS]{mvrnorm} is used, otherwise see function \link[mvtnorm]{rmvnorm}}

\item{seed}{the seed used for simulation (for reproducibility purpose)}
}
\value{
a list 
\item{data}{a expression matrix of \eqn{m\times n} where m is the number of
genes and n is the number of samples. }
\item{trt}{sample labels of length \code{n}, 1 for treatment and 0 for control.}
\item{go_term}{gene labels of length \code{m}, 1 for go_term genes and 0 otherwise.}
\item{sigma}{true covariance matrix upon which data is simulated.}
}
\description{
simulate normally distributed expression data with desired DE probabilities
for genes in the test set and for those not in the test set..
}
\examples{
t1 <- simulate_expression_data(size = 50, n_gene = 500, n_test = 100, 
                               prop = c(0.1, 0.1), de_mu = 2, de_sd = 1, 
                               rho1 = 0.1, rho2 = 0.05, rho3 = -0.05, 
                               data_gen_method = "chol", seed = 123)
}

